#SH-27 -(BSH)-Add confirmation step in scripting. #SH-27

+++ scripts/utilities/delete_summerhouse.sh
@@ -17,9 +17,19 @@
 #-----------------------------------------------------------------------------------------
 #
 
-## Stop DDEV and delete resources.
-yes |ddev delete -O 
-cd ../../..
-## Delete Summer House codebase and folders.
-rm -rf summerhouse 
-echo -e "   \e[1;4;31mYour local installation of Summer House was deleted.\e[0m"
+# Ask for confirmation.
+echo -e "   \e[1;4;31mThis script will execute a complete cleaning of the Summer House codebase in your system. \e[0m"
+echo -e "   \e[1;4;31mThis script has to be executed within the original folder: /scripts/utilities. \e[0m"
+read -p "Are you sure? [Y/y]" -n 1 -r
+echo    # move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  ## Stop DDEV and delete resources.
+  yes |ddev delete -O 
+  cd ../../..
+  ## Delete Summer House codebase and folders.
+  rm -rf summerhouse 
+  echo -e "   \e[1;4;31mYour local installation of Summer House was deleted.\e[0m"
+else
+  echo "The execution of the script will be stopped. Bye!";
+fi
